fully implemented wrmhl, made progress on moviemachine logic
Wrmhl is now playing nicely with Unity,
Arduino is fully responsive. Had to set
timeout time super high or else the controller would time out and stop functioning.

Created variables for each node, hooked them all up to each other. Traversing through the game is possible, but content still needs to be added. Must ask Adam for clarification on the nodes, as it seems that his flowchart may not make sense in the context of unity.